Refactor Index.ts to separate the creation of the app and the application start up
